First UserRest impl.

+++ src/main/java/ru/johnson/airblocknote/repository/NoteRepo.java
@@ -3,5 +3,5 @@ package ru.johnson.airblocknote.repository;
 import org.springframework.data.repository.CrudRepository;
 import ru.johnson.airblocknote.model.Note;
 
-public interface NoteRepo extends CrudRepository<Note, Long> {
+public interface NoteRepo extends CrudRepository<Note, Integer> {
 }
