+++ src/main/java/ru/johnson/airblocknote/repository/UserRepo.java
@@ -3,5 +3,5 @@ package ru.johnson.airblocknote.repository;
 import org.springframework.data.repository.CrudRepository;
 import ru.johnson.airblocknote.model.User;
 
-public interface UserRepo extends CrudRepository<User, Long> {
+public interface UserRepo extends CrudRepository<User, Integer> {
 }
